clean up and additinal comments
