+++ server/routes/profile.forms2.ts
@@ -10,6 +10,28 @@ profileForms2 route. Differs from termdb.profileFormScores:
   - Public role never sees site IDs (sites: []).
 */
 
+/*
+Builds termdbConfig.profileForms2Domain2PlotType for the Templates 2 picker:
+{ [cohort]: { [domainId]: [plotTypeName, ...] } }.
+
+The mapping is statically declared per cohort in the dataset config at
+plotConfigByCohort[cohort].profileForms2.domain2plotType — no SQL is needed.
+Empty inner submaps signal "cohort declared but no template data".
+
+Invoked from termdb.config.ts as a thin pass-through.
+*/
+export function getProfileForms2Domain2PlotType(ds: any) {
+	const tdb = ds.cohort?.termdb
+	if (!tdb?.plotConfigByCohort) return undefined
+	const out: Record<string, Record<string, string[]>> = {}
+	for (const cohortKey of Object.keys(tdb.plotConfigByCohort)) {
+		const cfg = tdb.plotConfigByCohort[cohortKey]?.profileForms2
+		if (!cfg?.options) continue
+		out[cohortKey] = cfg.domain2plotType || {}
+	}
+	return Object.keys(out).length ? out : undefined
+}
+
 export const api: RouteApi = {
 	endpoint: 'termdb/profileForms2Scores',
 	methods: {
